+++ derive-input/src/api/db.rs
@@ -1,66 +1,7 @@
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct SpacetimeDBTable {
-    pub singular_name: Box<str>,
-    pub visibility: SpacetimeDBTableVisibility,
-    pub multi_column_indices: Vec<Index>,
-    pub scheduled_reducer: Option<ScheduledReducer>,
-}
+pub mod table;
 
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub enum SpacetimeDBTableVisibility {
-    Public,
-    Private,
-}
+pub mod column;
 
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct Index {
-    pub name: Box<str>,
-    pub is_unique: bool,
-    pub index_type: IndexType,
-}
+pub mod index;
 
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub enum IndexType {
-    /// Available from `SpacetimeDBTable.multi_column_indices`
-    BTreeMultiColumn { columns: Vec<Box<str>> },
-    /// Available from `SpacetimeDBColumn.single_column_index`
-    BTreeSingleColumn { column: Box<str> },
-    /// Available from `SpacetimeDBColumn.single_column_index`
-    Direct { column: Box<str> },
-}
-
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct ScheduledReducer {
-    pub reducer_name: Box<str>,
-}
-
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct SpacetimeDBColumn {
-    pub is_primary_key: bool,
-    pub single_column_index: Option<Index>,
-    pub is_auto_inc: bool,
-}
+pub mod reducer;
