+++ derive-input/src/api/rust.rs
@@ -1,34 +1,5 @@
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct RustStruct {
-    pub visibility: RustVisibility,
-    pub name: Box<str>,
-}
+pub mod table;
 
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct RustField {
-    pub visibility: RustVisibility,
-    pub name: Box<str>,
-    pub type_name_or_path: Box<str>,
-}
+pub mod column;
 
-#[cfg_attr(feature = "clone", derive(Clone))]
-#[cfg_attr(feature = "debug", derive(Debug))]
-#[cfg_attr(feature = "partial-eq", derive(PartialEq))]
-#[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
-#[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub enum RustVisibility {
-    /// `pub`
-    Public,
-    /// `pub(crate)`, `pub(super)` or `pub(in path::to::module)`
-    Restricted(Box<str>),
-    /// Default
-    Private,
-}
+pub mod visibility;
