+++ derive-input/src/api/rust/table.rs
@@ -5,10 +5,7 @@ use crate::api::rust::visibility::RustVisibility;
 #[cfg_attr(feature = "partial-eq", derive(PartialEq))]
 #[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
 #[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct Setter {
-    pub method_visibility: RustVisibility,
-    pub method_name: Box<str>,
-    pub method_arg: Box<str>,
-    pub return_type: Box<str>,
-    pub method_impl: Box<str>,
+pub struct RustStruct {
+    pub visibility: RustVisibility,
+    pub name: Box<str>,
 }
