+++ derive-input/src/api/rust/visibility.rs
@@ -1,14 +1,13 @@
-use crate::api::rust::visibility::RustVisibility;
-
 #[cfg_attr(feature = "clone", derive(Clone))]
 #[cfg_attr(feature = "debug", derive(Debug))]
 #[cfg_attr(feature = "partial-eq", derive(PartialEq))]
 #[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
 #[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
-pub struct Setter {
-    pub method_visibility: RustVisibility,
-    pub method_name: Box<str>,
-    pub method_arg: Box<str>,
-    pub return_type: Box<str>,
-    pub method_impl: Box<str>,
+pub enum RustVisibility {
+    /// `pub`
+    Public,
+    /// `pub(crate)`, `pub(super)` or `pub(in path::to::module)`
+    Restricted(Box<str>),
+    /// Default
+    Private,
 }
