+++ derive-input/src/lib.rs
@@ -14,11 +14,11 @@ pub mod api {
     #[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
     #[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
     pub struct Table {
-        pub rust_struct: rust::RustStruct,
-        pub spacetimedb_table: db::SpacetimeDBTable,
+        pub rust_struct: rust::table::RustStruct,
+        pub spacetimedb_table: db::table::SpacetimeDBTable,
         pub spacetimedsl_table: dsl::table::SpacetimeDSLTable,
         pub columns: Vec<Column>,
-        pub spacetimedsl_methods: dsl::method::SpacetimeDSLTableMethods,
+        pub spacetimedsl_methods: dsl::table::SpacetimeDSLTableMethods,
     }
 
     impl Table {
@@ -43,10 +43,10 @@ pub mod api {
     #[cfg_attr(feature = "partial-ord", derive(PartialOrd))]
     #[cfg_attr(feature = "spacetime-type", derive(spacetimedb::SpacetimeType))]
     pub struct Column {
-        pub rust_field: rust::RustField,
-        pub spacetimedb_column: db::SpacetimeDBColumn,
+        pub rust_field: rust::column::RustField,
+        pub spacetimedb_column: db::column::SpacetimeDBColumn,
         pub spacetimedsl_column: dsl::column::SpacetimeDSLColumn,
-        pub spacetimedsl_methods: Option<dsl::method::SpacetimeDSLColumnMethods>,
+        pub spacetimedsl_methods: Option<dsl::column::SpacetimeDSLColumnMethods>,
     }
 }
